==> Notebooks/1. Pre_treatment.ipynb <==
# 1. Pre-treatment
Merges the clinical case data with MeSH term annotations and applies multiple filters to extract the final dataset of 10,000 English clinical cases, split into training (8,574), validation (1,072), and test (1,072) sets.


import os
import s3fs
import pandas as pd

os.chdir("/home/onyxia/projet_NLP")
os.getcwd()

## 1. Chargement des fichiers

full_df = pd.read_csv('medical_datasets/male_brain_tumor_dataset/cases.csv')
full_df.head()

full_df["pmcid"] = full_df["case_id"].apply(lambda x: x.split("_")[0])
full_df.head(2)

metadata = pd.read_parquet('medical_datasets/whole_multicare_dataset/metadata.parquet')
metadata.head()

# Création d'une DataFrame à partir des colonnes 'article_metadata' et 'article_id' de 'metadata'
df_metadata = pd.json_normalize(metadata['article_metadata'])
df_metadata['article_id'] = metadata['article_id']

df_metadata.head()

print('nombre de ligne avec keywords non vide : ' + str(df_metadata.keywords.notna().sum())) #Mettre pourcentage
print('nombre de ligne avec major_mesh_terms non vide : ' + str(df_metadata.major_mesh_terms.notna().sum()))

## 2. Filtering

* Suppression des articles avec plus d'un cas cliniques (case_amount > 1) du df_metadata

* Suppression des articles avec colonne major_mesh_terms vide

df_metadata_mesh = df_metadata.copy(deep=True)

df_metadata_mesh.dropna(subset='major_mesh_terms', inplace=True)
df_metadata_mesh.query("case_amount <= 1", inplace=True)

df_metadata_mesh.head(10)

df_metadata_mesh.shape

## 3. Left join of the full_df dataset into the metadata dataset
To apply our filters to **full_df** and retrieve a unified dataset containing both the `case_text` of the selected clinical cases and their associated `major MeSH terms`, we performed a left join by merging full_df into the metadata dataset using `pmcid` as the key.

df_merged = pd.merge(df_metadata_mesh, full_df, how='left')
df_merged.head(2)

columns = ['pmcid','title', 'gender', 'case_text', 'keywords', 'major_mesh_terms', 'mesh_terms', 'journal', 'doi']
columns2 = ['pmcid','title', 'keywords', 'major_mesh_terms']


# Set option to display all columns
pd.set_option('display.max_colwidth', None)
df = df_merged['case_text']
print(df.head(1))

df = df_merged['major_mesh_terms']
print(df.head(1))

## 4. Sauvegarde dataframe final 'df'

file = "df_final"
path = f'data/{file}.pkl'

if not os.path.exists(path):
    df.to_pickle(path)
else:
    print("Le fichier 'df_final.pkl' existe déjà. Aucune action effectuée.")
