==> Notebooks/2. Encoding_MeSH terms_in_Mesh_Code.ipynb <==
# 2. Encoding_MeSH_terms
Maps MeSH descriptors to their corresponding high-level chapter categories (C01–C26) to facilitate multi-label classification at the MeSH-C level.

## Classification Mesh 


Le MeSH (Medical Subject Headings) est un vocabulaire contrôlé utilisé pour organiser et indexer les articles médicaux. Il est structuré de manière hiérarchique en “arbres” (MeSH trees), où chaque concept est associé à un code unique appelé numéro d’arbre.

“Catégorie principale des Maladies nutritionnelles et métaboliques” est ainsi associé au code C14, où “C” représente la catégorie des maladies et “14” spécifie le groupe des maladies nutritionnelles et métaboliques. Cette organisation facilite la navigation du général au spécifique, aidant ainsi les chercheurs à trouver des informations précises.

Mais il est aussi d'être plus precis en selectionnant des maladies plus loin dans l'**arborescences**.

Par le Diabete de type 1 correspond au code MeSH C18.452.394.750.124 qui se décompose comme suit :

*   C : Catégorie des maladies
*   C18 : Catégorie principale des Maladies nutritionnelles et métaboliques.
*   C18.452 : Sous-catégorie des Maladies métaboliques.
*   C18.452.394 : Sous-sous-catégorie des Troubles du métabolisme du glucose.
*   C18.452.394.750 : Sous-sous-sous-catégorie du Diabète sucré.
*   C18.452.394.750.124 : Terme spécifique pour le Diabète sucré de type 1.

Cette hiérarchie reflète l’organisation des termes MeSH, permettant de situer chaque condition médicale dans un contexte plus large et de naviguer efficacement dans les informations médicales.

import pandas as pd
import numpy as np

import sys
import os

# Ajouter le dossier courant au PATH si ce n'est pas déjà fait
sys.path.append(os.path.abspath("."))

os.chdir("/home/onyxia/projet_NLP")
os.getcwd()

from Script import MeSH
print(MeSH.Hello())

Le dictionnaire est disponible au format XML à cet [URL](https://nlmpubs.nlm.nih.gov/projects/mesh/MESH_FILES/xmlmesh/?_gl=1*1vhz8wv*_ga*MTkwNTMyNjEyOC4xNzM2Njg2Mzcy*_ga_7147EPK006*MTczNzEwNTQ1MS4xMi4xLjE3MzcxMDU4OTMuMC4wLjA.*_ga_P1FPTH9PL4*MTczNzEwNTQ1MS4xNC4xLjE3MzcxMDU4OTMuMC4wLjA)

import requests

# URL et chemin local du fichier XML
url_fichier = "https://nlmpubs.nlm.nih.gov/projects/mesh/MESH_FILES/xmlmesh/desc2025.xml"
chemin_fichier = "data/desc2025.xml"  # Nom du fichier local

# Vérifier si le fichier existe déjà
if not os.path.exists(chemin_fichier):
    try:
        response = requests.get(url_fichier)
        response.raise_for_status()  # Vérifie si la requête a réussi

        with open(chemin_fichier, "wb") as fichier:
            fichier.write(response.content)

        print(f"Fichier téléchargé avec succès : {chemin_fichier}")
    except requests.RequestException as e:
        print(f"Erreur lors du téléchargement du fichier XML : {e}")
        exit(1)  # Arrête l'exécution en cas d'échec
else:
    print(f"Le fichier {chemin_fichier} existe déjà. Pas de téléchargement nécessaire.")

# Analyse du fichier téléchargé
racine = MeSH.chargement_arbre_mesh(chemin_fichier)

## Chargement données

df = pd.read_parquet('data/df_final.pkl')
df.head(3)

print(type(df.major_mesh_terms[4]))
print(type(MeSH.list_to_dict(df.major_mesh_terms[4])))
print(MeSH.list_to_dict(df.major_mesh_terms[4]))

df['dict_mesh'] = df.major_mesh_terms.apply(MeSH.list_to_dict)
df.head()

## Extraction Code MeSH

import multilabel_preprocessing as mp
from tqdm import tqdm
tqdm.pandas()  # Pour que apply() affiche une barre de progression

# J AI COPIE CE CODE SOUS """VERSION CACHED""" DANS LE FICHIER MESH --> POSSIBILITE D'UTILISER DIRECTEMENT CES FONCTIONS

from functools import lru_cache

@lru_cache(maxsize=None)
def rechercher_descripteur_cached(terme, racine):
    return MeSH.recherche_descripteur(racine, terme, format_reponse='numero_mesh')


def extraire_codes_disease_C_cached(dict_mesh_ligne, racine):
    """
    Version optimisée avec cache + rapide + propre.
    """
    codes = set()

    if not isinstance(dict_mesh_ligne, dict):
        return []

    for termes in dict_mesh_ligne.values():
        for terme in termes:
            list_mesh = rechercher_descripteur_cached(terme, racine)
            if list_mesh:
                for item in list_mesh:
                    if item.startswith("C") and len(item) > 2:
                        code = item[1:3]
                        if code.isdigit():
                            codes.add(code)

    return sorted(codes)

df2 = df.copy(deep=True)
df2.head(1)

#raise RuntimeError("⛔ Cette cellule ne doit pas être exécutée.")

# Application ligne par ligne avec racine et objet MeSH passés comme arguments
df2["codes_mesh_extraits"] = df2["dict_mesh"].progress_apply(lambda x: extraire_codes_disease_C_cached(x, racine))

df2["target"] = df2["codes_mesh_extraits"].apply(lambda x: mp.vectorizer(x))

#raise RuntimeError("⛔ Cette cellule ne doit pas être exécutée.")
df2.to_pickle("data/df_target_V2.pkl")

from sklearn.model_selection import train_test_split

df = pd.read_pickle("data/df_target_V2.pkl")

# Split the dataset, create X (features) and y (target), print the size
data_train, data_test = train_test_split(df ,test_size=0.10, random_state=42)

print ('train size',data_train.shape)
print ('test size', data_test.shape)



data_train.to_pickle("data/df_train.pkl")
data_test.to_pickle("data/df_test.pkl")
