==> Notebooks/4. Naive_Bayesian_classifier.ipynb <==
# Naive Bayes Classifier

1. **Réaliser un TF-IDF** sur l'ensemble des `cases_texte` du `df_final`.
2. **Tenter de prédire** la colonne `major_mesh_terms` à partir du TF-IDF.

## Import data

import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from tqdm import tqdm  # Barre de progression pour Jupyter
pd.set_option('display.max_colwidth', None)
import s3fs


AWS = False

if os.path.exists("../data/df_train.pkl"):
    data_train = pd.read_pickle('../data/df_train.pkl')
    print(data_train.shape)

if os.path.exists("../data/df_test.pkl"):
    data_test = pd.read_pickle('../data/df_test.pkl')
    print(data_test.shape)

os.environ["AWS_ACCESS_KEY_ID"] = 'TX7EAFIX4BX0YG40M0HS'
os.environ["AWS_SECRET_ACCESS_KEY"] = '6u6aMhddxRy+JkoE1+Gb53Rf+PqPync9xgyO6q3W'
os.environ["AWS_SESSION_TOKEN"] = 'eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJhY2Nlc3NLZXkiOiJUWDdFQUZJWDRCWDBZRzQwTTBIUyIsImFsbG93ZWQtb3JpZ2lucyI6WyIqIl0sImF1ZCI6WyJtaW5pby1kYXRhbm9kZSIsIm9ueXhpYSIsImFjY291bnQiXSwiYXV0aF90aW1lIjoxNzQyNTczNTU4LCJhenAiOiJvbnl4aWEiLCJlbWFpbCI6InF1ZW50aW4ubWFycmV0QGVuc2FlLmZyIiwiZW1haWxfdmVyaWZpZWQiOnRydWUsImV4cCI6MTc0MzYxMDc3OCwiZmFtaWx5X25hbWUiOiJNYXJyZXQiLCJnaXZlbl9uYW1lIjoiUXVlbnRpbiIsImdyb3VwcyI6WyJVU0VSX09OWVhJQSJdLCJpYXQiOjE3NDMwMDU5NzgsImlzcyI6Imh0dHBzOi8vYXV0aC5sYWIuc3NwY2xvdWQuZnIvYXV0aC9yZWFsbXMvc3NwY2xvdWQiLCJqdGkiOiJjYWQ4MDk4OS05ZTY4LTQ1MzAtODlhYS0wM2UxZTliZTg3ZGQiLCJsb2NhbGUiOiJmciIsIm5hbWUiOiJRdWVudGluIE1hcnJldCIsInBvbGljeSI6InN0c29ubHkiLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJxdWVudGluMTk5OSIsInJlYWxtX2FjY2VzcyI6eyJyb2xlcyI6WyJvZmZsaW5lX2FjY2VzcyIsInVtYV9hdXRob3JpemF0aW9uIiwiZGVmYXVsdC1yb2xlcy1zc3BjbG91ZCJdfSwicmVzb3VyY2VfYWNjZXNzIjp7ImFjY291bnQiOnsicm9sZXMiOlsibWFuYWdlLWFjY291bnQiLCJtYW5hZ2UtYWNjb3VudC1saW5rcyIsInZpZXctcHJvZmlsZSJdfX0sInJvbGVzIjpbIm9mZmxpbmVfYWNjZXNzIiwidW1hX2F1dGhvcml6YXRpb24iLCJkZWZhdWx0LXJvbGVzLXNzcGNsb3VkIl0sInNjb3BlIjoib3BlbmlkIHByb2ZpbGUgZ3JvdXBzIGVtYWlsIiwic2lkIjoiZTlmMWI0YzgtNTYwNC00M2UwLTk2NDUtZjM0M2Q2Y2U0NDVmIiwic3ViIjoiOGM1NjkxMjYtMGI2Mi00MGI0LTg4ZDAtMWFhYzgyMTJiOGE5IiwidHlwIjoiQmVhcmVyIn0.Gl-pro2vlaMpQZiX9x1b9ngp10WD8yH8_KIi8VqofdmnaJ67cdOyIEQ_2NJ-V-XZAjSCC6fC8L7sD8POngUs6g'
os.environ["AWS_DEFAULT_REGION"] = 'us-east-1'
fs = s3fs.S3FileSystem(
    client_kwargs={'endpoint_url': 'https://'+'minio.lab.sspcloud.fr'},
    key = os.environ["AWS_ACCESS_KEY_ID"], 
    secret = os.environ["AWS_SECRET_ACCESS_KEY"], 
    token = os.environ["AWS_SESSION_TOKEN"])

print(repr(os.getenv("AWS_ACCESS_KEY_ID")))
print(repr(os.getenv("AWS_SECRET_ACCESS_KEY")))

raise RuntimeError("⛔ Cette cellule ne doit pas être exécutée.")

#Create filesystem object
S3_ENDPOINT_URL = "https://" + os.environ["AWS_S3_ENDPOINT"]
fs = s3fs.S3FileSystem(client_kwargs={'endpoint_url': S3_ENDPOINT_URL})

BUCKET_OUT = "s3://quentin1999/Data_Projet_NLP"
FILE_KEY_OUT_S3 = "df_final"
FILE_PATH_OUT_S3 = BUCKET_OUT + "/" + FILE_KEY_OUT_S3

with fs.open(FILE_PATH_OUT_S3, 'w') as file_out:
    df.to_csv(file_out)

if AWS == True:
    fs.ls("s3://quentin1999/Data_Projet_NLP")

if AWS == True:
    if not os.path.exists("../data/df_target_V2.pkl"):
        BUCKET = "s3://quentin1999/Data_Projet_NLP"
        FILE_KEY_S3 = "df_target_V2.pkl"
        FILE_PATH_S3 = BUCKET + "/" + FILE_KEY_S3

        with fs.open(FILE_PATH_S3, mode="rb") as file_in:
            df = pd.read_pickle(file_in)

        print("✅ Fichier chargé avec succès, shape :", df.shape)

        df.to_pickle("data/df_target_V2.pkl")

        print("✅ Fichier sauvegardé avec succès dans le dossier data")

## Naive Bayes Classifier

### Bag-of-word representation

from sklearn.model_selection import train_test_split

print ('train size',data_train.shape)
print ('test size', data_test.shape)

# create features X and target y
X_train = data_train.case_text
X_test = data_test.case_text

y_train = data_train.target
y_test = data_test.target

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

vectorizer = CountVectorizer(max_features=1000)
#vectorizer = TfidfVectorizer(ngram_range=(1, 2), max_features=3000)
# fit on train data
vectorizer.fit(X_train)

# Apply it on train and dev data
X_train_counts = vectorizer.transform(X_train)
X_test_counts = vectorizer.transform(X_test)

vectorizer

### OneVsRestClassifier avec CountVectorizer

from sklearn.linear_model import LogisticRegression
from sklearn.multiclass import OneVsRestClassifier

# Initiate the model
# Définition du modèle multi-label simple avec Logistic Regression
model = OneVsRestClassifier(LogisticRegression(solver='liblinear'))

# Train
model.fit(X_train_counts, np.vstack(y_train.values))
y_pred = (model.predict_proba(X_test_counts) > 0.5).astype(int)

from sklearn.metrics import f1_score, hamming_loss, classification_report
import numpy as np

# 1. Conversion correcte de y_test (stocké sous forme de vecteurs dans une colonne)
y_test_array = np.vstack(y_test.values)

# 2. Calcul des métriques
f1_micro = f1_score(y_test_array, y_pred, average='micro')
f1_macro = f1_score(y_test_array, y_pred, average='macro')
hamming = hamming_loss(y_test_array, y_pred)

# 3. Affichage des métriques globales
print(f"F1 micro       : {f1_micro:.4f}")
print(f"F1 macro       : {f1_macro:.4f}")
print(f"Hamming loss   : {hamming:.4f}")

# 4. Classification report
print("\nClassification Report :")
print(classification_report(y_test_array, y_pred, zero_division=0))

### OneVsRestClassifier avec TF-IDF ngram = 2

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

vectorizer = TfidfVectorizer(ngram_range=(1, 2), max_features=5000)
# fit on train data
vectorizer.fit(X_train)

# Apply it on train and dev data
X_train_counts = vectorizer.transform(X_train)
X_test_counts = vectorizer.transform(X_test)

# Initiate the model
# Définition du modèle multi-label simple avec Logistic Regression
model = OneVsRestClassifier(LogisticRegression(solver='liblinear'))

# Train
model.fit(X_train_counts, np.vstack(y_train.values))
y_pred = (model.predict_proba(X_test_counts) > 0.25).astype(int)

# 1. Conversion correcte de y_test (stocké sous forme de vecteurs dans une colonne)
y_test_array = np.vstack(y_test.values)

# 2. Calcul des métriques
f1_micro = f1_score(y_test_array, y_pred, average='micro')
f1_macro = f1_score(y_test_array, y_pred, average='macro')
hamming = hamming_loss(y_test_array, y_pred)

# 3. Affichage des métriques globales
print(f"F1 micro       : {f1_micro:.4f}")
print(f"F1 macro       : {f1_macro:.4f}")
print(f"Hamming loss   : {hamming:.4f}")

# 4. Classification report
print("\nClassification Report :")
print(classification_report(y_test_array, y_pred, zero_division=0))
