==> Notebooks/8. Translation.ipynb <==
# 8. Translation into French of English clinical cases

Extraction of clinical cases from data_train & data_test to be translated in French
* This notebook were just used to select a small portion of clinical cases from the english train & test sets
* The translation was then made using Google translate (included in Google Sheet)

import os
import pandas as pd
import numpy as np

os.chdir("/home/onyxia/work/projet_NLP")
os.getcwd()

data_train = pd.read_pickle("data/df_train.pkl")
data_test = pd.read_pickle("data/df_test.pkl")

print(data_train.shape)
print(data_test.shape)

sample_train = data_train.sample(167, random_state=42)
sample_test = data_test.sample(300, random_state=42)

if not os.path.exists("data/Translated/train_trans_fr.csv"):
    sample_train.to_csv("data/Translated/train_trans_fr.csv")
    print("ficher sample_train créé avec succès")
if not os.path.exists("data/Translated/test_trans_fr.csv"):
    sample_test.to_csv("data/Translated/test_trans_fr.csv")
    print("ficher sample_test créé avec succès")

print(sample_train.shape)
print(sample_test.shape)
